requirements.txt
APScheduler==3.6.3
asgiref==3.5.2
atomicwrites==1.4.1
attrs==22.1.0
cachetools==4.2.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==2.0.12
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==42.0.1
defusedxml==0.8.0rc2
Django==3.2
django-filter==23.5
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
exceptiongroup==1.1.3
flake8==5.0.4
idna==3.6
iniconfig==1.1.1
itypes==1.2.0
Jinja2==3.1.3
MarkupSafe==2.1.4
mccabe==0.7.0
oauthlib==3.2.2
packaging==21.3
pep8-naming==0.13.3
Pillow==9.3.0
pluggy==0.13.1
prettytable==3.8.0
py==1.11.0
pyaes==1.6.1
pycodestyle==2.9.1
pycparser==2.21
pydocstyle==6.3.0
pyflakes==2.5.0
PyJWT==2.1.0
pyparsing==3.0.9
Pyrogram==2.0.106
PySocks==1.7.1
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
python-dotenv==1.0.0
python-telegram-bot==13.7
python3-openid==3.2.0
pytils==0.4.1
pytz==2022.6
requests==2.26.0
requests-oauthlib==1.3.1
six==1.16.0
snowballstemmer==2.2.0
social-auth-app-django==4.0.0
social-auth-core==4.4.2
sqlparse==0.4.3
toml==0.10.2
tomli==2.0.1
tornado==6.4
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
uritemplate==4.1.1
urllib3==1.26.18
wcwidth==0.2.6
webcolors==1.13
